Align Finance schema assertion types

+++ tests/Depot.Tests/FinanceBudgetingTests.cs
@@ -23,7 +23,7 @@ public sealed class FinanceBudgetingTests
 			using var connection = new SqliteConnection($"Data Source={path}");
 			connection.Open();
 
-			Assert.Equal(FinanceInventoryAccountingSchemaMigration.CurrentVersion, Scalar(connection, "SELECT Version FROM DepotFeatureVersions WHERE Name='Finance';"));
+			Assert.Equal((long)FinanceInventoryAccountingSchemaMigration.CurrentVersion, Scalar(connection, "SELECT Version FROM DepotFeatureVersions WHERE Name='Finance';"));
 			Assert.Equal(1L, Scalar(connection, "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='FinanceBudgetVersions';"));
 			Assert.Equal(1L, Scalar(connection, "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='FinanceBudgetLines';"));
 			Assert.Equal(1L, Scalar(connection, "SELECT COUNT(*) FROM sqlite_master WHERE type='index' AND name='IX_FinanceBudgetVersions_List';"));
